Finished part 5a.

+++ bloglist-backend/controllers/blogs.js
@@ -1,13 +1,27 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1, id: 1})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {  
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {  
+  const body = request.body
+
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
 
   if (blog.likes == undefined) {
     blog.likes = 0
@@ -15,6 +29,8 @@ blogsRouter.post('/', async (request, response) => {
   
   if (blog.title != undefined && blog.url != undefined) {
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
   }else{
     response.status(400).end()
@@ -37,9 +53,19 @@ blogsRouter.put('/:id', async (request, response) => {
   response.status(200).end()
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (user._id.toString() === blog.user.toString()){
+    console.log(`Blog ${request.params.id} successfully deleted.`)
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    return response.status(401).json({ error: 'cannot delete blog posted by another user' })
+  }
 })
 
 module.exports = blogsRouter
